Move Simplifier helper functions to Simplifier helpers files
In order for a downstream project to access the following functions,
move them from OMRSimplifierHandlers.cpp into OMRSimplifierHelpers:
-foldUByteConstant()
-foldCharConstant()
-setCCAddSigned()
-setCCAddUnsigned()
-setCCSubSigned()
-setCCSubUnsigned()
-setCCOr()
-calculateSignedCC()
-calculateUnsignedCC()
